requirements.txt
# python==3.9.7
zhipuai==2.1.4
# antlr4-python3-runtime==4.13.1
antlr4-tools==0.2.1
antlr4-python3-runtime==4.13.1
rank_bm25==0.2.2
pytz==2021.3
jieba==0.42.1
flask==3.0.3
python-docx==0.8.11
httpx==0.27.2
openai==1.51.2
transformers==4.45.2
anthropic==0.36.0
modelscope==1.18.1
accelerate==1.4.0
tiktoken==0.9.0
python-dotenv==1.1.0